Consolidate client/server files
