+++ packages/core/package.json
@@ -37,6 +37,7 @@
     "ethers": "^4.0.30",
     "express": "^4.17.1",
     "ganache-cli": "^6.4.4",
+    "jayson": "^3.0.2",
     "level": "^5.0.1",
     "uuid": "^3.3.2",
     "web3": "^1.0.0-beta.55",
