+++ packages/core/src/app/common/index.ts
@@ -1,4 +1,3 @@
 export * from './eth'
 export * from './db'
-export * from './net'
 export * from './utils'
