--- packages/core/src/app/common/net/index.ts
@@ -1,2 +1,0 @@
-export * from './rpc'
-export * from './transport'
