--- packages/core/src/app/common/net/rpc/index.ts
@@ -1,1 +1,0 @@
-export * from './json-rpc'
