--- packages/core/src/app/common/net/transport/http/index.ts
@@ -1,2 +1,0 @@
-export * from './axios-http-client'
-export * from './express-http-server'
