--- packages/core/src/app/common/net/transport/index.ts
@@ -1,1 +1,0 @@
-export * from './http'
