+++ packages/core/src/app/index.ts
@@ -1,2 +1,3 @@
 export * from './client'
 export * from './common'
+export * from './transport'
