+++ packages/core/src/app/transport/client/axios-http-client.ts
@@ -2,7 +2,7 @@
 import axios, { AxiosRequestConfig, AxiosResponse } from 'axios'
 
 /* Internal Imports */
-import { HttpClient } from '../../../../../interfaces'
+import { HttpClient } from 'src/interfaces'
 
 /**
  * HTTP client that uses the axios client library.
