+++ packages/core/src/app/transport/client/index.ts
@@ -1,4 +1,3 @@
+export * from './axios-http-client'
 export * from './json-rpc-client'
-export * from './json-rpc-server'
-export * from './json-rpc-errors'
 export * from './json-rpc-http-adapter'
