+++ packages/core/src/interfaces/index.ts
@@ -2,3 +2,4 @@ export * from './client'
 export * from './common'
 export * from './core'
 export * from './client'
+export * from './transport'
