+++ packages/core/src/interfaces/transport/index.ts
@@ -1,0 +1,4 @@
+export interface ServerBackend {
+  app: any
+  listen(): Promise<void>
+}
