+++ packages/core/test/app/transport/json-rpc-server.spec.ts
@@ -5,9 +5,8 @@ import debug from 'debug'
 const log = debug('test:info:state-manager')
 
 /* Internal Imports */
-import { JsonRpcClient, JsonRpcHttpAdapter, JsonRpcServer } from '../../../src/app/common/net/rpc'
-import { AxiosHttpClient } from '../../../src/app/common/net/transport'
-import { HttpRequest, HttpResponse } from '../../../src/interfaces'
+import { JsonRpcClient, JsonRpcHttpAdapter, AxiosHttpClient,  JsonRpcServer } from 'src/app'
+import { HttpRequest, HttpResponse } from 'src/interfaces'
 
 describe.only('JSON RPC Server', () => {
   describe('Server', () => {
@@ -17,7 +16,7 @@ describe.only('JSON RPC Server', () => {
       }
       const server = new JsonRpcServer({
         test: testFn
-      }, 3000, 'localhost')
+      }, 'localhost', 3000)
       await server.listen()
       const client = new JsonRpcClient<HttpRequest, HttpResponse>(new JsonRpcHttpAdapter(), new AxiosHttpClient('http://127.0.0.1:3000'))
       await client.handle('test', 'test123')
